requirements.txt
pandas
streamlit
python-dotenv
streamlit_option_menu 
streamlit_chat
africastalking
openai
seaborn
requests
matplotlib
numpy
folium
google-generativeai
streamlit-folium
folium
streamlit-stl
openrouteservice
plotly
geopy
langchain
langchain-community
langchain-core
langchain-google-genai
google-generativeai
bcrypt
captcha
python-dotenv
tqdm
chromadb
faiss-cpu
pypdf
